add chainID config parameter for bsc
 move ChildChainTransaction structure to stand alone file
